Updated Backend
All get functions are now more flexible with regards to from and to arguments and now supports passing date and POSITX classes along with characters.

Bugfix: Bitmart didnt return anything on the weekly candle. This has been fixed.

All APIs now return 100 pips as intended when either FROM and TO is NULL.