rlang-dependency is gone, all tests have passed!
This partially closes https://github.com/serkor1/cryptoQuotes/issues/10

All rlang-dependency have been removed from all charts. There are currently two tasks left for this:

1. write an asssert-statement in the remove_bounds-function, as this was using rlang. It is currently just commented out - the FUN.R-functions needs to be revised and its usecase double-checked before its worth spending more time on these functions.

2. asdd_event-function has been commented out. No other reason than being lazy at this point. this function needs to be revised as it was too restrictive in its structure. So it needs some thorough work.

NOTE: the package startupmessages have been migrated to cli::inform(). This comes with the downside that the colors in the terminal are not dynamic. But this serves no utiility other than a wow-factor; so this wont be dealt with anytime soon.

-----------------------------

Other elements in this commit:

1. Removed redundant files; chart_indicators.R
2. Restructured chart_ma; it now has a helper-function. This function is a specific helper,
and therefore it havent been migrated to the chart_helpers (see 3.)
3. Created a new helper script: chart_helpers, this script should reduce the code needed
to build the charts.
4. paletteer-package dependency has been removed; all colors are now extracted manually.

Steps taken to make code more efficient:

As of now the code has been modularized, and the indicators calculated via TTR::foo() has been wrapped in the functions indicators and pull. These functions needs decent documentation at a later point.