Commit closes https://github.com/serkor1/cryptoQuotes/issues/9
The "In (function (...)  : 'tzone' attributes are inconsistent" error was caused by the comparision in max(date, coerce()) for Binance endpoints in LS-ratioes. This was also the case for the remaining functions where there were a hard day limit on returned values. These have been rearranged to support the comparisions.

The testthat functions have been updated to expect_no_condition to capture these kinds of errors in advance.