Bug-fix in infer_interval
The infer_interval()-function would throw an error on Tradfi OHLC candles as these follows holidays, and are closed in weekends. This is mitigated by the find_mode-function which is essentially just the statistical mode. It does introduce some overhead, but this is minimal around 600 microseconds as per latest tests.