Major Changes
The line-chart have been renamed to pline to avoid interferrence with the line-function from the stats-package. New and more concise examples for LSR, MAs, FGI and pline() have been made to avoid noisy examples.

The MA example is a single example using a few of the MA-functions. It  is expected that the user can fill in the blanks from here.

Most of the documentation has been streamlined. There is one more round of streamlining, and the proofreading before it is finished.

From here:

1. Create a custom TA function
2. Improve on pkgdown website

Then push to development. From there squash to main - then we publish to CRAN.