Updated Documentation and Minor Bugfix
The pkgdown website now has a different, and shorter, landing page. Wether this is a good idea will be clear in time.

All articles has been renamed, and the _pkgdown.yml is more low-level and detailed than before. Its a rabbithole once you start this - but it looks good nonetheless.

Bugfix: There was a bug where if **to** wasnt defined in the get_quote()-function the returned subset would be returned in a different timezone that made merging a bit awkward across endpoints.