rOpenSci Compliance Update
The pkgcheck::pkgcheck()-function was flagging the chart()-function for using global assignment operators. To avoid the use of these applicable lapply-functions have been replaced with Map and/or Reduce.

Part of the helper code have been simplified.

Bugfix:

ohlc and kline charts were displaying the borders on hover. These have been removed.

Next Commit:

Indicators are not colorblind friendly. This will be changed in the next commit.