Binance, Assertand Fetch Power-Ups :arrow_up: :arrow_up:
Assert:

The assert function has been reworked so it only evaluates all statements if any of the values are FALSE. The update is twice as fast as before in cases where all statements are TRUE.

Fetch:

The fetch()-function is now less defensive and uses only a single try-catch statement.

GET:

The GET()-function is now less defensive

Binance:

All possible parameters have been vectorized and/or simplified where possible.

General:

All ifelse()-functions have been replaced with if () else as these are faster.
For cases where more than two if statements are needed, a switch-function is used instead (with default values)